Web Development MCQ Quiz
All questions require answer. For every correct answers, user will get 10 points and for every wrong answer user loses 5 points and at the end of quiz, marks is displayed and based upon that review is given to user and taken back to user's choice whether he/she wants to continue it or not.